Continue restructure and modularization
--break out modules
--create routes
--maintain boilerplate file

This commit doesn't function due to issues in relative imports, but is close
