+++ .idea/misc.xml
@@ -3,5 +3,5 @@
   <component name="Black">
     <option name="sdkName" value="Python 3.9 (customer_analysis)" />
   </component>
-  <component name="ProjectRootManager" version="2" project-jdk-name="Python 3.12 (customer_analysis)" project-jdk-type="Python SDK" />
+  <component name="ProjectRootManager" version="2" project-jdk-name="Python 3.9 (customer_analysis)" project-jdk-type="Python SDK" />
 </project>
