--- app/utils/file_utils_2.py
@@ -1,7 +1,0 @@
-def upload_file():
-    # Code to handle file upload
-    pass
-
-def download_file(data):
-    # Code to handle file download
-    pass
