--- to_sort/Response.csv
@@ -1,60 +1,0 @@
-,customer_data_model_fields,guessed_matching_column_names,confidence_percentage
-2,last_name,LastName,95
-3,email,EmailAddress,95
-1,first_name,FirstName,95
-0,contact_id,ClientID,90
-4,phone,PhoneNumber,90
-15,address_line1,ShippingAddress,85
-12,phone_number,PhoneNumber,85
-8,companies,Company,80
-11,customer_id,ClientID,80
-41,churn_risk,,0
-44,occupation,,0
-43,customer_notes,,0
-42,account_status,,0
-39,referral_count,,0
-40,feedback_score,,0
-46,income_level,,0
-38,referral_code,,0
-37,preferred_language,,0
-36,last_login_date,,0
-45,industry,,0
-50,household_size,,0
-47,education_level,,0
-48,marital_status,,0
-49,children_count,,0
-34,customer_segment,,0
-51,annual_spend,,0
-52,favorite_product,,0
-53,preferred_payment_method,,0
-54,account_manager,,0
-55,social_media_handle,,0
-56,customer_satisfaction_score,,0
-57,preferred_contact_time,,0
-35,marketing_source,,0
-29,communication_preference,,0
-33,opt_in_sms,,0
-19,postal_code,,0
-5,created_at,,0
-6,updated_at,,0
-7,properties,,0
-9,deals,,0
-10,tickets,,0
-13,date_of_birth,,0
-14,gender,,0
-16,address_line2,,0
-17,city,,0
-18,state,,0
-20,country,,0
-32,opt_in_email,,0
-21,join_date,,0
-22,last_purchase_strategy,,0
-23,total_spend,,0
-24,average_order_value,,0
-25,order_count,,0
-26,is_loyalty_member,,0
-27,loyalty_points,,0
-28,preferred_store,,0
-30,subscription_status,,0
-31,has_purchased,,0
-58,extra_fields,,0
